(Re)move legacy ical4android code and tests
