+++ core/src/main/kotlin/at/bitfire/davdroid/db/migration/AutoMigration16.kt
@@ -8,7 +8,7 @@ import androidx.room.ProvidedAutoMigrationSpec
 import androidx.room.RenameColumn
 import androidx.room.migration.AutoMigrationSpec
 import androidx.sqlite.db.SupportSQLiteDatabase
-import at.bitfire.ical4android.ICalendar
+import at.bitfire.synctools.icalendar.timezoneDefToTzId
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -29,7 +29,7 @@ class AutoMigration16 @Inject constructor(): AutoMigrationSpec {
             while (cursor.moveToNext()) {
                 val id: Long = cursor.getLong(0)
                 val timezoneDef: String = cursor.getString(1) ?: continue
-                val timezoneId = ICalendar.timezoneDefToTzId(timezoneDef)
+                val timezoneId = timezoneDefToTzId(timezoneDef)
                 db.execSQL("UPDATE collection SET timezoneId=? WHERE id=?", arrayOf<Any?>(timezoneId, id))
             }
         }
