+++ core/src/main/kotlin/at/bitfire/davdroid/settings/migration/AccountSettingsMigration12.kt
@@ -12,7 +12,7 @@ import android.provider.CalendarContract
 import android.util.Base64
 import androidx.core.content.ContextCompat
 import androidx.core.content.contentValuesOf
-import at.bitfire.ical4android.UnknownProperty
+import at.bitfire.synctools.storage.UnknownProperty
 import at.bitfire.synctools.storage.calendar.EventsContract
 import at.techbee.jtx.JtxContract.asSyncAdapter
 import dagger.Binds
