+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/jtx/JtxBatchOperationTest.kt
@@ -8,9 +8,9 @@ import android.accounts.Account
 import android.content.ContentProviderClient
 import androidx.core.content.contentValuesOf
 import androidx.test.platform.app.InstrumentationRegistry
-import at.bitfire.ical4android.util.MiscUtils.asSyncAdapter
 import at.bitfire.synctools.storage.BatchOperation
 import at.bitfire.synctools.storage.TaskProvider
+import at.bitfire.synctools.storage.asSyncAdapter
 import at.bitfire.synctools.test.BuildConfig
 import at.bitfire.synctools.test.GrantPermissionOrSkipRule
 import at.techbee.jtx.JtxContract
