+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/calendar/builder/AllDayBuilder.kt
@@ -6,7 +6,7 @@ package at.bitfire.synctools.mapping.calendar.builder
 
 import android.content.Entity
 import android.provider.CalendarContract.Events
-import at.bitfire.ical4android.util.DateUtils
+import at.bitfire.synctools.icalendar.DateUtils
 import at.bitfire.synctools.icalendar.dtStart
 import net.fortuna.ical4j.model.component.VEvent
 import java.time.temporal.Temporal
