+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/calendar/builder/ColorBuilder.kt
@@ -8,7 +8,7 @@ import android.content.Entity
 import android.provider.CalendarContract.Colors
 import android.provider.CalendarContract.Events
 import androidx.annotation.VisibleForTesting
-import at.bitfire.ical4android.util.MiscUtils.asSyncAdapter
+import at.bitfire.synctools.storage.asSyncAdapter
 import at.bitfire.synctools.storage.calendar.AndroidCalendar
 import net.fortuna.ical4j.model.component.VEvent
 import net.fortuna.ical4j.model.property.Color
