+++ synctools/src/main/kotlin/at/bitfire/synctools/mapping/calendar/handler/DurationHandler.kt
@@ -6,9 +6,9 @@ package at.bitfire.synctools.mapping.calendar.handler
 
 import android.content.Entity
 import android.provider.CalendarContract.Events
-import at.bitfire.ical4android.util.TimeApiExtensions.abs
 import at.bitfire.synctools.icalendar.plusAssign
 import at.bitfire.synctools.util.AndroidTimeUtils
+import at.bitfire.synctools.util.TimeApiExtensions.abs
 import net.fortuna.ical4j.model.component.VEvent
 import net.fortuna.ical4j.model.property.DtEnd
 import java.time.Instant
