+++ synctools/src/test/kotlin/at/bitfire/synctools/mapping/calendar/builder/UnknownPropertiesBuilderTest.kt
@@ -8,8 +8,8 @@ import android.content.ContentValues
 import android.content.Entity
 import android.provider.CalendarContract.ExtendedProperties
 import androidx.core.content.contentValuesOf
-import at.bitfire.ical4android.UnknownProperty
 import at.bitfire.synctools.icalendar.propertyListOf
+import at.bitfire.synctools.storage.UnknownProperty
 import at.bitfire.synctools.test.assertContentValuesEqual
 import net.fortuna.ical4j.model.component.VEvent
 import net.fortuna.ical4j.model.parameter.XParameter
