+++ synctools/src/test/kotlin/at/bitfire/synctools/mapping/tasks/builder/UnknownPropertiesBuilderTest.kt
@@ -8,8 +8,8 @@ import android.content.ContentValues
 import android.content.Entity
 import android.net.Uri
 import androidx.core.content.contentValuesOf
-import at.bitfire.ical4android.UnknownProperty
 import at.bitfire.synctools.mapping.tasks.VToDoUtil
+import at.bitfire.synctools.storage.UnknownProperty
 import at.bitfire.synctools.storage.tasks.DmfsTaskList
 import at.bitfire.synctools.storage.tasks.DmfsTasksContract.UNKNOWN_PROPERTY_DATA
 import at.bitfire.synctools.test.assertContentValuesEqual
