+++ synctools/src/test/kotlin/at/bitfire/synctools/mapping/tasks/handler/UnknownPropertiesHandlerTest.kt
@@ -8,7 +8,7 @@ import android.content.ContentValues
 import android.content.Entity
 import android.net.Uri
 import androidx.core.content.contentValuesOf
-import at.bitfire.ical4android.UnknownProperty
+import at.bitfire.synctools.storage.UnknownProperty
 import at.bitfire.synctools.storage.tasks.DmfsTasksContract.UNKNOWN_PROPERTY_DATA
 import net.fortuna.ical4j.model.Parameter
 import net.fortuna.ical4j.model.component.VToDo
